Split out EditPlayer
