+++ src/screens/ConfigureScreen.js
@@ -1,13 +1,12 @@
 import React, { useState } from 'react';
-import { Platform, Text, View, ScrollView, StyleSheet, TextInput, Image, Dimensions } from 'react-native';
+import { Platform, Text, View, StyleSheet, Image, Dimensions } from 'react-native';
 import { useSelector, useDispatch } from 'react-redux';
-import { setPlayerName, newGame, addPlayer, removePlayer } from '../../redux/CurrentGameActions';
+import { newGame, addPlayer } from '../../redux/CurrentGameActions';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
-import { ListItem, Avatar, Badge, Icon, Card, Input, Button } from 'react-native-elements'
+import { Icon, Button } from 'react-native-elements'
 import Animated from 'react-native-reanimated';
 import BottomSheet from 'reanimated-bottom-sheet';
-import { DefaultTheme, DarkTheme, useTheme } from "@react-navigation/native";
-import { getContrastRatio } from 'colorsheet';
+import EditPlayer from '../components/EditPlayer';
 
 const appJson = require('../../app.json');
 
@@ -19,11 +18,6 @@ const ConfigureScreen = () => {
     const players = useSelector(state => state.currentGame.players);
     const dispatch = useDispatch();
 
-    const setPlayerNameHandler = (index, name) => {
-        dispatch(setPlayerName(index, name));
-        setPlayerWasAdded(false)
-    }
-
     const newGameHandler = () => {
         dispatch(newGame());
         setIsNewGame(true);
@@ -32,10 +26,6 @@ const ConfigureScreen = () => {
     const addPlayerHandler = () => {
         dispatch(addPlayer('Player ' + (players.length + 1)));
         setPlayerWasAdded(true)
-    }
-
-    const removePlayerHandler = (index) => {
-        dispatch(removePlayer(index));
     }
 
     const sheetRef = React.useRef(null);
@@ -90,7 +80,8 @@ const ConfigureScreen = () => {
                 <View style={{ margin: 10, }}>
                     <Button
                         icon={<Icon name="refresh" color="white" />}
-                        title="New Game" onPress={newGameHandler} />
+                        title="New Game"
+                        onPress={newGameHandler} />
                     {isNewGame &&
                         <Text style={{ textAlign: 'center', paddingTop: 10, color: '#eee' }}>
                             Scores have been reset!
@@ -103,66 +94,16 @@ const ConfigureScreen = () => {
                         Players
                     </Text>
 
-                    <Text style={{ fontSize: 18, color: '#0a84ff', marginTop: 20, textAlign: 'right' }}>Edit</Text>
+                    {/* <Text style={{ fontSize: 18, color: '#0a84ff', marginTop: 20, textAlign: 'right' }}>Edit</Text> */}
                 </View>
 
                 {players.map((player, index) => (
-                    <View style={styles.playerContainer} key={player.uuid}>
-                        <Text style={{
-                            fontSize: 35,
-                            padding: 5,
-                            fontWeight: "bold",
-                            color: "#0a84ff"
-                        }}>{index + 1}</Text>
-
-                        <View style={{
-                            width: 30,
-                            height: 30,
-                            borderRadius: 25,
-                            borderWidth: 1,
-                            borderColor: '#eee',
-                            padding: 5,
-                            marginHorizontal: 5
-                        }}
-                            onTouchStart={promptColor}
-                            backgroundColor={"#" + palette[index]}>
-                            <Icon
-                                name="edit"
-                                size={20}
-                                color={getContrastRatio('#' + palette[index], '#000').number > 7 ? "black" : "white"}
-                            ></Icon>
-                        </View>
-
-                        <Input style={styles.input}
-                            containerStyle={{ flex: 1 }}
-                            defaultValue={index == players.length - 1 && playerWasAdded ? null : player.name}
-                            autoFocus={index == players.length - 1 && playerWasAdded}
-                            placeholder={'Player ' + (index + 1)}
-                            selectTextOnFocus={true}
-                            renderErrorMessage={false}
-                            // label={"Player  " + (index + 1) + " Name"}
-                            onEndEditing={(e) => {
-                                if (e.nativeEvent.text == "") {
-                                    setPlayerNameHandler(index, 'Player ' + (index + 1));
-                                }
-                            }}
-                            maxLength={15}
-                            onChangeText={(text) => setPlayerNameHandler(index, text)} />
-
-                        {index > 0 &&
-                            <Icon name="delete" color="#ff375f" onPress={() => removePlayerHandler(index)}></Icon>
-                        }
-                    </View>
+                    <EditPlayer player={player} index={index} promptColor={promptColor} setPlayerWasAdded={setPlayerWasAdded} playerWasAdded={playerWasAdded} />
                 ))}
 
                 <View style={{ margin: 10 }}>
                     <Button title="Add Player"
-                        icon={
-                            <Icon
-                                name="add"
-                                color="white"
-                            />
-                        }
+                        icon={<Icon name="add" color="white" />}
                         disabled={players.length >= 8}
                         onPress={addPlayerHandler} />
                 </View>
@@ -199,16 +140,6 @@ const styles = StyleSheet.create({
     text: {
         fontSize: 18,
         margin: 15,
-        color: '#eee',
-    },
-    playerContainer: {
-        margin: 10,
-        marginVertical: 5,
-        justifyContent: 'flex-start',
-        alignItems: 'center',
-        flexDirection: 'row',
-    },
-    input: {
         color: '#eee',
     },
 });
